Separate darwin syscalls for mmap
Using the `unix` package by default for other systems instead of `syscall`
