+++ y/mmap.go
@@ -1,4 +1,4 @@
-// +build !windows,!darwin
+// +build !windows
 
 /*
  * Copyright 2017 Dgraph Labs, Inc. and Contributors
@@ -20,32 +20,22 @@ package y
 
 import (
 	"os"
-
-	"golang.org/x/sys/unix"
 )
 
 // Mmap uses the mmap system call to memory-map a file. If writable is true,
 // memory protection of the pages is set so that they may be written to as well.
-func mmap(fd *os.File, writable bool, size int64) ([]byte, error) {
-	mtype := unix.PROT_READ
-	if writable {
-		mtype |= unix.PROT_WRITE
-	}
-	return unix.Mmap(int(fd.Fd()), 0, int(size), mtype, unix.MAP_SHARED)
+func Mmap(fd *os.File, writable bool, size int64) ([]byte, error) {
+	return mmap(fd, writable, size)
 }
 
 // Munmap unmaps a previously mapped slice.
-func munmap(b []byte) error {
-	return unix.Munmap(b)
+func Munmap(b []byte) error {
+	return munmap(b)
 }
 
 // Madvise uses the madvise system call to give advise about the use of memory
 // when using a slice that is memory-mapped to a file. Set the readahead flag to
 // false if page references are expected in random order.
-func madvise(b []byte, readahead bool) error {
-	flags := unix.MADV_NORMAL
-	if !readahead {
-		flags = unix.MADV_RANDOM
-	}
-	return unix.Madvise(b, flags)
+func Madvise(b []byte, readahead bool) error {
+	return madvise(b, readahead)
 }
